+++ packages/contact-center/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.28.0-ccwidgets.97](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.96...v1.28.0-ccwidgets.97) (2025-07-28)
+
+
+### Bug Fixes
+
+* **widgets:** fixed-consult-transfer-popover-transient-issue ([#506](https://github.com/webex/widgets/issues/506)) ([ae646bd](https://github.com/webex/widgets/commit/ae646bde44ee4fbc5383cfc456c22d2f91b47d20))
+
+# [1.28.0-ccwidgets.96](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.95...v1.28.0-ccwidgets.96) (2025-07-24)
+
+
+### Features
+
+* **cc-widgets:** UI Automation for Basic Task Controls ([#497](https://github.com/webex/widgets/issues/497)) ([6aa1c90](https://github.com/webex/widgets/commit/6aa1c906340aa03ff6f2d43f8e57d3064e327c9e))
+
 # [1.28.0-ccwidgets.95](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.94...v1.28.0-ccwidgets.95) (2025-07-24)